test sucess M***********

+++ src/lib.rs
@@ -1,7 +1,5 @@
 // lib.rs
-use std::os::raw::c_int;
-
 #[no_mangle]
-pub extern "C" fn add_numbers(a: c_int, b: c_int) -> c_int {
+pub extern "C" fn add_numbers(a: i64, b: i64) -> i64 {
     a + b
 }
